+++ gix/src/worktree/mod.rs
@@ -120,18 +120,25 @@ pub mod excludes {
         /// Configure a file-system cache checking if files below the repository are excluded.
         ///
         /// This takes into consideration all the usual repository configuration.
-        // TODO: test, provide higher-level interface that is much easier to use and doesn't panic.
+        ///
+        /// `index` may be used to obtain `.gitignore` files directly from the index under certain conditions.
+        // TODO: test, provide higher-level interface that is much easier to use and doesn't panic when accessing entries
+        //       by non-relative path.
+        // TODO: `index` might be so special (given the conditions we are talking about) that it's better obtained internally
+        //        so the caller won't have to care.
+        // TODO: global files like `~/.gitignore` seem to be missing here, but we need a way to control if these should be loaded.
+        //       probably that needs another permission in the repo options or a custom config variable. The latter is easiest to manage.
         pub fn excludes(
             &self,
             index: &gix_index::State,
             overrides: Option<gix_attributes::MatchGroup<gix_attributes::Ignore>>,
         ) -> Result<gix_worktree::fs::Cache, Error> {
             let repo = self.parent;
-            let case = repo
-                .config
-                .ignore_case
-                .then_some(gix_glob::pattern::Case::Fold)
-                .unwrap_or_default();
+            let case = if repo.config.ignore_case {
+                gix_glob::pattern::Case::Fold
+            } else {
+                gix_glob::pattern::Case::Sensitive
+            };
             let mut buf = Vec::with_capacity(512);
             let excludes_file = match repo.config.excludes_file().transpose()? {
                 Some(user_path) => Some(user_path),
